test(#1491): split the ordering cases into their own module (campsite rule)
